+++ Release/Runtime/Zenject/LocalizationKeyAttribute.cs
@@ -1,0 +1,8 @@
+using UnityEngine;
+
+namespace Playdarium.Localization.Runtime.Zenject
+{
+	public class LocalizationKeyAttribute : PropertyAttribute
+	{
+	}
+}